backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
beautifulsoup4
transformers
torch
newspaper3k
celery
redis
